feat(todoapp): add confirmation dialog for delete action
Adds a `window.confirm` dialog before executing the task deletion in the
`TodoApp` component. This prevents accidental data loss when the user
clicks the delete button.

+++ examples/TodoApp.js
@@ -77,7 +77,9 @@ class TodoApp extends gia.Component {
 
 		if (deleteBtn) {
 			const id = deleteBtn.getAttribute('data-id');
-			this.deleteTask(id);
+			if (window.confirm('Are you sure you want to delete this task?')) {
+				this.deleteTask(id);
+			}
 		}
 	}
